Add example of 'overriding' methods using compose
At runtime this behaves as we would hope, as the overriding
functions are what in the end gets called, and is somewhat easier
to do with compose than with normal inheritance. However, there is
still the issue that at compile time the overriding and overridden
signature are both still considered to be valid by the compiler.

Trying to cast to the desired interface at the end doesn't seem
to be working, or at least not as it was being done in cellSelection.ts